Create Planet Order Lookup.r
The objective of this code is to retrieve the order ID and its associated order name. This is useful for when the need to download a specific file from the Planet GUI arises, especially if the account has hundreds/thousands of orders in its history. When you find the desired order ID, simply append it to the URL https://www.planet.com/account/#/orders/